Use the common webpack configuration
