+++ package.json
@@ -13,9 +13,9 @@
 	"author": "John Molakvoæ (skjnldsv) <skjnldsv@protonmail.com>",
 	"license": "AGPL-3.0",
 	"scripts": {
-		"dev": "webpack --config webpack.dev.js",
-		"watch": "webpack --progress --watch --config webpack.dev.js",
-		"build": "NODE_ENV=production webpack --progress --config webpack.prod.js",
+		"build": "NODE_ENV=production webpack --progress",
+		"dev": "NODE_ENV=development webpack --progress",
+		"watch": "NODE_ENV=development webpack --progress --watch",
 		"l10n:extract": "node build/extract-l10n.js",
 		"lint": "eslint --ext .js,.vue src",
 		"lint:fix": "eslint --ext .js,.vue src --fix",
@@ -78,39 +78,23 @@
 		"@nextcloud/browserslist-config": "^2.1.0",
 		"@nextcloud/eslint-config": "^7.0.2",
 		"@nextcloud/stylelint-config": "^2.1.2",
+		"@nextcloud/webpack-vue-config": "^5.0.0",
 		"@vue/test-utils": "^1.0.4",
 		"@vue/vue2-jest": "^27.0.0-alpha.4",
 		"babel-core": "^7.0.0-bridge.0",
 		"babel-jest": "^27.0.1",
-		"babel-loader": "^8.1.0",
-		"babel-loader-exclude-node-modules-except": "^1.0.3",
-		"css-loader": "^5.0.0",
 		"cypress": "^9.5.3",
 		"cypress-visual-regression": "^1.5.0",
 		"eslint-plugin-cypress": "^2.11.1",
-		"eslint-webpack-plugin": "^3.1.1",
-		"file-loader": "^6.0.0",
 		"gettext-extractor": "^3.5.2",
 		"gettext-parser": "^5.0.0",
-		"glob": "^7.1.6",
 		"jest": "^27.0.4",
 		"jest-environment-jsdom": "27.5.1",
 		"jest-serializer-vue": "^2.0.2",
 		"jest-transform-stub": "^2.0.0",
-		"node-sass": "^7.0.1",
-		"raw-loader": "^4.0.1",
 		"resolve-url-loader": "^5.0.0",
 		"sanitize-filename": "^1.6.3",
-		"sass-loader": "^12.6.0",
-		"stylelint-webpack-plugin": "^3.2.0",
-		"url-loader": "^4.1.0",
-		"vue-eslint-parser": "^8.0.1",
-		"vue-loader": "^15.9.1",
 		"vue-styleguidist": "^4.44.27",
-		"vue-template-compiler": "^2.6.14",
-		"webpack": "^5.71.0",
-		"webpack-cli": "^4.0.0",
-		"webpack-merge": "^5.0.8",
 		"webpack-node-externals": "^3.0.0"
 	},
 	"jest": {
