+++ styleguide.config.js
@@ -1,6 +1,6 @@
 const path = require('path')
 const { merge } = require('webpack-merge')
-const webpackConfig = require('./webpack.dev.js')
+const webpackConfig = require('./webpack.config.js')
 
 const newConfig = Object.assign({}, webpackConfig, {
 	externals: {},
