--- webpack.dev.js
@@ -1,12 +1,0 @@
-const { merge } = require('webpack-merge');
-const common = require('./webpack.common.js');
-
-module.exports = merge(common, {
-	mode: 'development',
-	devServer: {
-		historyApiFallback: true,
-		noInfo: true,
-		overlay: true
-	},
-	devtool: 'source-map',
-})
