--- webpack.prod.js
@@ -1,7 +1,0 @@
-const { merge } = require('webpack-merge')
-const common = require('./webpack.common.js')
-
-module.exports = merge(common, {
-	mode: 'production',
-	devtool: 'source-map'
-})
